[1607]: Move the assertions into their own package and split them into smaller files. This prevents a circular dependency where hold -> testutil -> app -> hold.
