+++ x/hold/genesis_test.go
@@ -9,7 +9,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/provenance-io/provenance/testutil"
+	"github.com/provenance-io/provenance/testutil/assertions"
 )
 
 func TestDefaultGenesisState(t *testing.T) {
@@ -154,7 +154,7 @@ func TestGenesisState_Validate(t *testing.T) {
 				err = tc.genState.Validate()
 			}
 			require.NotPanics(t, testFunc, "Validate()")
-			testutil.AssertErrorContents(t, err, tc.expErr, "Validate()")
+			assertions.AssertErrorContents(t, err, tc.expErr, "Validate()")
 		})
 	}
 }
